rag/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
chromadb>=0.4.0
pydantic>=2.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
GitPython>=3.1.0